Monitoring: Move Query Browser components and CSS to monitoring/
Create new components/monitoring/ directory and move in the Query
Browser related components and CSS.

Fix QueryBrowserProps type.
