+++ frontend/public/components/graphs/graph-loader.ts
@@ -1,5 +1,4 @@
 export { Bar } from './bar';
 export { Gauge } from './gauge';
 export { Area } from './area';
-export { QueryBrowser } from './query-browser';
 export { Donut } from './donut';
