+++ frontend/public/style.scss
@@ -55,6 +55,7 @@
 @import "components/global-notification";
 @import "components/list";
 @import "components/monitoring";
+@import "components/monitoring/monitoring";
 @import "components/nav-title";
 @import "components/node-ip-list";
 @import "components/RBAC/rbac";
